# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.BaseResult:
    properties:
      error:
        type: string
      result: {}
    type: object
  dto.CreateSubscription:
    properties:
      price:
        type: integer
      service_name:
        type: string
      user_id:
        type: string
    type: object
  dto.ErrorResult:
    properties:
      error:
        type: string
    type: object
  dto.UpdateSubscription:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  title: Test Task
  version: "1.0"
paths:
  /subscriptions/:
    post:
      consumes:
      - application/json
      parameters:
      - description: Body for creating a new subscription
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSubscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ErrorResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResult'
      summary: Creating new subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: subscription id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResult'
      summary: Delete subscription
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      parameters:
      - description: subscription id
        in: path
        name: id
        required: true
        type: string
      - description: Body for updating a new subscription
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSubscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResult'
      summary: Update subscription
      tags:
      - subscriptions
  /subscriptions/{user_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: filter by user id
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.BaseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResult'
      summary: Get all subscriptions
      tags:
      - subscriptions
  /subscriptions/all:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResult'
      summary: Get all subscriptions
      tags:
      - subscriptions
  /subscriptions/calculate:
    get:
      consumes:
      - application/json
      description: '''start_date'' and ''end_date'' queries is required. ''user_id''
        and ''service_name'' is optional'
      parameters:
      - description: 'Start of subscription. format: ''YYYY-MM-DD'' (required)'
        in: query
        name: start_date
        required: true
        type: string
      - description: 'End of subscription. format: ''YYYY-MM-DD'' (required)'
        in: query
        name: end_date
        required: true
        type: string
      - description: Filter by user id (optional)
        in: query
        name: user_id
        type: string
      - description: Filter by service name (optional)
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResult'
      summary: Get sum of subscriptions
      tags:
      - subscriptions
schemes:
- http
swagger: "2.0"